completed the di

+++ mlsWrapper/src/main/java/com/uney/android/mls/mlswrapper/WrapperConfiguration.kt
@@ -2,7 +2,7 @@ package com.uney.android.mls.mlswrapper
 
 interface WrapperConfiguration {
     val casBaseUrl: String
-    val casTimeout: Int
+    val casTimeout: Long
     val uniqueClientId: String
     val cryptoStoragePath: String
     val onAccessTokenRequested: () -> String
